+++ nenupy/undysputed/dynspec.py
@@ -2,10 +2,73 @@
 # -*- coding: utf-8 -*-
 
 
-"""
+r"""
     *******
     Dynspec
     *******
+
+    **UnDySPuTeD**
+
+    *UnDySPuTeD* (stands for Unified Dynamic Spectrum Pulsar and
+    Time Domain receiver) is the receiver of the NenuFAR
+    beamformer mode, fed by the (up-to-)96 core Mini-Arrays (2
+    polarizations) from the *LANewBa* backend. 
+
+    **DynSpec data**
+
+    The raw data flow from *LANewBa* consists of 195312.5 pairs
+    of complex X and Y values per second per beamlet. These data
+    are downsampled in channels per subband (of 195.3125 kHz)
+    numbered from 16 to 2048 channels, ``fftlen``, (to achieve a
+    frequency resolution of :math:`\delta \nu = 12 - 0.1\, \rm{kHz}`
+    respectively). After computation of cross and auto-correlations,
+    the data are downsampled again in time, integrating from 4 to 1024
+    spectra, ``nfft2int`` (implying a time resolution :math:`195312.5^{-1} \times \rm{fftlen} \times \rm{fft2int}`,
+    :math:`\delta t = 0.3 - 83.9\, \rm{ms}`).
+
+    .. seealso::
+        `DynSpec data product <https://nenufar.obs-nancay.fr/en/astronomer/#data-products>`_
+
+    **DynSpec data files**
+    
+    Each NenuFAR/*UnDySPuTeD*/DynSpec observation results in the
+    production of several proprietary formatted files (``'*.spectra'``),
+    each corresponding to an individual lane of the *UnDySPuTeD* receiver.
+    Depending on the observation configuration, the bandwidth and/or
+    the different observed beams (i.e., beamforming in different sky
+    directions) can be spread accross these files.
+
+    The class :class:`~nenupy.undysputed.dynspec.Dynspec` offers
+    the possibility to read and analyze all these observation files
+    at once, through an API aimed at minimizing user interaction
+    over complicated settings.
+    
+    For the purpose of the following 'how-to' guide, a 1h-observation
+    of the famous pulsar `PSR B1919+21 <https://en.wikipedia.org/wiki/PSR_B1919%2B21>`_
+    is analyzed. :mod:`~nenupy.undysputed.dynspec.Dynspec` needs
+    first to be imported, as well as :mod:`~astropy.units` in order
+    to use physcial units (to avoid mistakes).
+    Attribute :attr:`~nenupy.undysputed.dynspec.Dynspec.lanefiles`
+    is filled with a `list` of all the available lane files in order to create
+    ``ds``, an instance of :class:`~nenupy.undysputed.dynspec.Dynspec`:
+
+    >>> from nenupy.undysputed import Dynspec
+    >>> import astropy.units as u
+    >>> ds = Dynspec(
+            lanefiles=[
+                'B1919+21_TRACKING_20200214_090835_0.spectra',
+                'B1919+21_TRACKING_20200214_090835_1.spectra'
+            ]
+        )
+
+    .. note::
+        `nenupy` logger could be activated at wish to enhance the verbosity:
+
+        >>> import logging
+        >>> logging.getLogger('nenupy').setLevel(logging.INFO)
+    
+    **Observation properties**
+
 """
 
 
@@ -21,7 +84,7 @@ __all__ = [
 ]
 
 
-from os.path import isfile, abspath, join, dirname
+from os.path import isfile, abspath, join, dirname, basename
 import numpy as np
 import astropy.units as u
 from astropy.time import Time
@@ -29,6 +92,7 @@ from dask.array import from_array
 from dask.diagnostics import ProgressBar
 
 from nenupy.beamlet import SData
+from nenupy.astro import dispersion_delay
 
 import logging
 log = logging.getLogger(__name__)
@@ -48,9 +112,9 @@ class _Lane(object):
         self.df = None
         self.fft0 = None
         self.fft1 = None
-        self.beam = None
+        self.beam_arr = None
         self.chan = None
-        self.unix = None
+        # self.unix = None
         self.lanefile = lanefile
 
 
@@ -67,6 +131,10 @@ class _Lane(object):
         if not isfile(l):
             raise FileNotFoundError(
                 f'{l} not found.'
+            )
+        if not l.endswith('spectra'):
+            raise ValueError(
+                'Wrong file type, expected *.spectra'
             )
         self._lanefile = l
         self._load()
@@ -210,7 +278,7 @@ class _Lane(object):
         # Store data in appropriated attributes
         self.fft0 = data['fft0']
         self.fft1 = data['fft1']
-        self.beam = data['beam']
+        self.beam_arr = data['beam'][0].compute() # asummed same for all time step
         self.chan = data['channel']
         ntb, nfb = data['lane'].shape # time blocks, freq blocks
         # self.unix = np.arange(ntb, dtype='float64')
@@ -235,6 +303,15 @@ class _Lane(object):
         self._freqs *= self.df.to(u.Hz).value
         self._freqs += self.chan[0, :][:, np.newaxis] * self.subband_width.value # assumed constant over time
         self._freqs = self._freqs.ravel()
+        # Beam indices
+        unique_b = np.unique(self.beam_arr)
+        self.beam_idx = {}
+        for b in unique_b:
+            b_idx = np.where(self.beam_arr == b)[0]
+            self.beam_idx[str(b)] = (
+                b_idx[0] * self.fftlen, # freq start of beam
+                (b_idx[-1] + 1) * self.fftlen # freq stop of beam
+            )
         return
 
 
@@ -311,6 +388,7 @@ class _Lane(object):
 # ============================================================= #
 class Dynspec(object):
     """
+        .. versionadded:: 1.1.0
     """
 
     def __init__(self, lanefiles=[]):
@@ -320,20 +398,58 @@ class Dynspec(object):
         self.dispersion_measure = None
         self.rebin_dt = None
         self.rebin_df = None
+        self.jump_correction = True
 
 
     # --------------------------------------------------------- #
     # --------------------- Getter/Setter --------------------- #
     @property
     def lanefiles(self):
+        """ UnDySPuTeD time-frequency files input (i.e. one file
+            per lane) that are stored in the attribute 
+            :attr:`~nenupy.undysputed.dynspec.Dynspec.lanes` as
+            :class:`~nenupy.undysputed.dynspec._Lane` instances.
+
+            Subsequent selections in time, frequency, beam and
+            data operation processes will be run on these files
+            while querying only those required. It means that any
+            file listed in input is not necessarily read, and
+            therefore having more files in input than needed
+            does not increase memory usage or computing time.
+
+            :setter: `list` of ``'*.spectra'`` files.
+
+            :getter: `list` of ``'*.spectra'`` files.
+            
+            :type: `list`
+
+            :Example:
+                >>> from nenupy.undysputed import Dynspec
+                >>> ds = Dynspec(
+                        lanefiles=['lane_0.spectra', 'lane_0.spectra']
+                    )
+                >>> ds.lanefiles
+                ['/path/to/lane_0.spectra', '/path/to/lane_0.spectra']
+
+            .. warning::
+                Files are checked to belong to the same
+                observation based on their prefix.
+    
+            .. versionadded:: 1.1.0
         """
-        """
-        return self._lanefiles
+        return list(self._lanefiles)
     @lanefiles.setter
     def lanefiles(self, l):
         if not isinstance(l, list):
             l = [l]
-        l = map(abspath, l)
+        l = list(map(abspath, l))
+        # Check that all the input files belong to the same obs
+        names = list(map(basename,l))
+        obsname = [ni[::-1].split('_', 1)[1] for ni in names]
+        if obsname != obsname[::-1]:
+            raise ValueError(
+                'Input files seem not to belong to the same observation.'
+            )
         for li in map(abspath, l):
             if not isfile(li):
                 raise FileNotFoundError(
@@ -344,12 +460,14 @@ class Dynspec(object):
                 self.lanes.append(
                     _Lane(li)
                 )
+        self._lanefiles = l
         return
 
 
     @property
     def time_range(self):
         """
+            .. versionadded:: 1.1.0
         """
         if not hasattr(self, '_time_range'):
             return [self.tmin.unix, self.tmax.unix]
@@ -380,6 +498,7 @@ class Dynspec(object):
     @property
     def freq_range(self):
         """
+            .. versionadded:: 1.1.0
         """
         if not hasattr(self, '_freq_range'):
             return [self.fmin.to(u.Hz), self.fmax.to(u.Hz)]
@@ -408,8 +527,32 @@ class Dynspec(object):
 
 
     @property
+    def beam(self):
+        """
+            .. versionadded:: 1.1.0
+        """
+        return self._beam
+    @beam.setter
+    def beam(self, b):
+        if not isinstance(b, (int, np.integer)):
+            raise TypeError(
+                'beam should be an integer.'
+            )
+        if b not in self.beams:
+            raise ValueError(
+                f'Available beam indices are {self.beams}.'
+            )
+        self._beam = b
+        log.info(
+            f'Beam index set: {b}.'
+        )
+        return
+
+
+    @property
     def dispersion_measure(self):
         """
+            .. versionadded:: 1.1.0
         """
         return self._dispersion_measure
     @dispersion_measure.setter
@@ -428,6 +571,7 @@ class Dynspec(object):
     @property
     def rebin_dt(self):
         """
+            .. versionadded:: 1.1.0
         """
         return self._rebin_dt
     @rebin_dt.setter
@@ -446,6 +590,7 @@ class Dynspec(object):
     @property
     def rebin_df(self):
         """
+            .. versionadded:: 1.1.0
         """
         return self._rebin_df
     @rebin_df.setter
@@ -464,6 +609,7 @@ class Dynspec(object):
     @property
     def tmin(self):
         """
+            .. versionadded:: 1.1.0
         """
         return min(li.tmin for li in self.lanes)
 
@@ -471,6 +617,7 @@ class Dynspec(object):
     @property
     def tmax(self):
         """
+            .. versionadded:: 1.1.0
         """
         return max(li.tmax for li in self.lanes)
 
@@ -478,6 +625,7 @@ class Dynspec(object):
     @property
     def fmin(self):
         """
+            .. versionadded:: 1.1.0
         """
         fm = min(li.fmin for li in self.lanes)
         return fm.compute().to(u.MHz)
@@ -486,17 +634,36 @@ class Dynspec(object):
     @property
     def fmax(self):
         """
+            .. versionadded:: 1.1.0
         """
         fm = max(li.fmax for li in self.lanes)
         return fm.compute().to(u.MHz)
-    
+
+
+    @property
+    def beams(self):
+        """
+            .. versionadded:: 1.1.0
+        """
+        un_beams = []
+        for li in self.lanes:
+            lane_beams = [int(lbi) for lbi in li.beam_idx.keys()]
+            un_beams += lane_beams
+        return np.unique(un_beams)
+
 
     # --------------------------------------------------------- #
     # ------------------------ Methods ------------------------ #
     def get(self, stokes='I'):
         """
+            .. versionadded:: 1.1.0
         """
         for li in self.lanes:
+            try:
+                beam_start, beam_stop = li.beam_idx[str(self.beam)]
+            except KeyError:
+                # No beam 'self.beam' found on this lane
+                continue
             data = li.get_stokes(stokes)
             # Time selection
             # way more efficient to compute indices than masking
@@ -506,24 +673,32 @@ class Dynspec(object):
             tmax_idx = np.argmin(
                 np.abs(li._times - self.time_range[1])
             ).compute()
-            # Freq selection
+            # Freq/beam selection
             # way more efficient to compute indices than masking
             fmin_idx = np.argmin(
-                np.abs(li._freqs - self.freq_range[0].value)
+                np.abs(li._freqs[beam_start:beam_stop] - self.freq_range[0].value)
             ).compute()
             fmax_idx = np.argmin(
-                np.abs(li._freqs - self.freq_range[1].value)
+                np.abs(li._freqs[beam_start:beam_stop] - self.freq_range[1].value)
             ).compute()
+            if (fmin_idx - fmax_idx) == 0:
+                # No data selected on this lane
+                continue
             log.info(
                 f'Retrieving data selection from lane {li.lane_index}...'
             )
             # High-rate data selection
             with ProgressBar():
-                data = data[
+                data = data[:, beam_start:beam_stop][
                     tmin_idx:tmax_idx,
                     fmin_idx:fmax_idx
                 ].compute()
-            selfreqs = li._freqs[fmin_idx:fmax_idx].compute()*u.Hz
+            selfreqs = li._freqs[beam_start:beam_stop][fmin_idx:fmax_idx].compute()*u.Hz
+            # Correct 6 min jumps
+            data = self._correct_jumps(
+                data=data,
+                dt=li.dt 
+            )
             # Dedispersion if FRB or Pulsar
             data = self._dedisperse(
                 data=data,
@@ -553,10 +728,12 @@ class Dynspec(object):
                 spec = spec & sd
         return spec
 
+
     # --------------------------------------------------------- #
     # ----------------------- Internal ------------------------ #
     def _dedisperse(self, data, freqs, dt):
         """
+            .. versionadded:: 1.1.0
         """
         if not (self.dispersion_measure is None):
             log.info(
@@ -567,10 +744,13 @@ class Dynspec(object):
                     'Problem with frequency axis.'
                 )
             dm = self.dispersion_measure.value # pc/cm^3
-            fmhz = freqs.to(u.MHz).value # MHz
-            fmaxmhz = self.freq_range[1].to(u.MHz).value
-            delays = 4140 * dm / (fmhz**2) * u.s # sec
-            delays -= 4140 * dm / (fmaxmhz**2) *u.s # start of dynspec
+            delays = dispersion_delay(
+                freq=freqs,
+                dm=self.dispersion_measure)
+            delays -= dispersion_delay( # relative delays
+                freq=self.freq_range[1],
+                dm=self.dispersion_measure
+            )
             cell_delays = np.round(delays/dt).astype(int)
             for i in range(freqs.size):
                 data[:, i] = np.roll(data[:, i], -cell_delays[i])
@@ -584,16 +764,20 @@ class Dynspec(object):
 
     def _rebin(self, data, times, freqs, dt, df):
         """
-
+            .. versionadded:: 1.1.0
         """
         ntimes_i, nfreqs_i = data.shape
         if not (self.rebin_dt is None):
             # Rebin in time
+            tbins = int(np.floor(self.rebin_dt/dt))
             log.info(
-                'Time-averaging...'
+                f'Time-averaging {tbins} spectra, dt={tbins*dt}...'
             )
-            ntimes = int(np.floor(ntimes_i/(self.rebin_dt/dt)))
+            ntimes = int(np.floor(ntimes_i/tbins))
             tleftover = ntimes_i % ntimes
+            log.info(
+                f'Last {tleftover} spectra are left over for time-averaging.'
+            )
             data = data[:-tleftover if tleftover != 0 else ntimes_i, :].reshape(
                 (ntimes, int((ntimes_i - tleftover)/ntimes), nfreqs_i)
             )
@@ -602,24 +786,21 @@ class Dynspec(object):
             )
             data = np.mean(data, axis=1)
             times = np.mean(times, axis=1)
-            times = Time(
-                times,
-                format='unix',
-                precision=7
-            )
             ntimes_i, nfreqs_i = data.shape
             log.info(
                 'Data are time-averaged.'
             )
         if not (self.rebin_df is None):
             # Rebin in frequency
+            fbins = int(np.floor(self.rebin_df/df))
             log.info(
-                'Frequency-averaging...'
+                f'Frequency-averaging {fbins} channels: df={fbins*df}...'
             )
-            data = data
-            nfreqs = int(np.floor(nfreqs_i/(self.rebin_df/df)))
+            nfreqs = int(np.floor(nfreqs_i/fbins))
             fleftover = nfreqs_i % nfreqs
-            #fleftover = fleftover if fleftover != 0 else nfreqs_i
+            log.info(
+                f'Last {fleftover} channels are left over for frequency-averaging.'
+            )
             data = data[:, :-fleftover if fleftover != 0 else nfreqs_i].reshape(
                 (ntimes_i, nfreqs, int((nfreqs_i - fleftover)/nfreqs))
             )
@@ -631,7 +812,52 @@ class Dynspec(object):
             log.info(
                 'Data are frequency-averaged.'
             )
+        times = Time(
+            times,
+            format='unix',
+            precision=7
+        )
         return data, times, freqs
 
+
+    def _correct_jumps(self, data, dt):
+        """
+        """
+        if not self.jump_correction:
+            return data
+        # Work on cleanest profile
+        #std_freq = np.std(data, axis=0)
+        #profile = data[:, np.argmin(std_freq)].copy()
+        profile = np.median(data, axis=1)
+        # detect jumps
+        dt = dt.to(u.s).value
+        deriv = np.gradient(profile, dt)
+        low_threshold = np.nanmedian(deriv) - 8 * np.nanstd(deriv)
+        lower_thr_indices = np.where(deriv < low_threshold)[0] # detection
+        # Sort jump indices
+        jump_idx = [0]
+        for idx in lower_thr_indices:
+            if idx > jump_idx[-1] + 10:
+                jump_idx.append(idx)
+        jump_idx.append(data.shape[0] - 1)
+        # Make sure there are no > 6min gaps
+        sixmin = 6*60.000000 + dt
+        for idx in range(1, len(jump_idx)):
+            delta_t = dt * (jump_idx[idx] - jump_idx[idx-1])
+            if delta_t > sixmin:
+                missing_idx = int(np.round(sixmin/dt))
+                jump_idx.insert(idx, missing_idx + jump_idx[idx-1])
+        # flatten
+        boundaries = list(map(list, zip(jump_idx, jump_idx[1:]))) 
+        for i, boundary in enumerate(boundaries):
+            idi, idf = boundary
+            med_data_freq = profile[idi:idf]
+            data[idi:idf, :] /= med_data_freq[:, np.newaxis]
+            if i == 0:
+                data[idi:idf, :] *= np.median(med_data_freq)
+            else:
+                data[idi:idf, :] *= previous_endpoint
+            previous_endpoint = np.median(data[idf-1, :])
+        return data
 # ============================================================= #